[MLIR] Fix Windows build and update docs
* Fix compilation issues under Windows.
* Extend `Design Principles` section in documentation.
* Enable build of experimental MLIR-based compiler by default.
  * For linux only, Windows has issues on pre-commit CI (need to update VS).

+++ src/experimental/vpux_compiler/include/vpux/compiler/core/dims_order.hpp
@@ -90,9 +90,9 @@ public:
 public:
     bool hasDim(Dim d) const;
 
-    int32_t dimPos(Dim d) const;
+    size_t dimPos(Dim d) const;
 
-    Dim dimAt(int32_t pos) const;
+    Dim dimAt(size_t pos) const;
 
 public:
     MemDim toMemDim(Dim d) const;
